+++ stroom-query-api/src/main/java/stroom/query/api/v2/ExpressionItem.java
@@ -105,7 +105,7 @@ public abstract class ExpressionItem implements Serializable {
 
         }
 
-        public Builder(final boolean enabled) {
+        public Builder(final Boolean enabled) {
             this.enabled = enabled;
         }