+++ stroom-query-api/src/main/java/stroom/query/api/v2/ExpressionOperator.java
@@ -21,8 +21,17 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import stroom.util.shared.HasDisplayValue;
 
-import javax.xml.bind.annotation.*;
-import java.util.*;
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlElementWrapper;
+import javax.xml.bind.annotation.XmlElements;
+import javax.xml.bind.annotation.XmlType;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.List;
+import java.util.Objects;
 
 @JsonPropertyOrder({"op", "children"})
 @XmlType(name = "ExpressionOperator", propOrder = {"op", "children"})
@@ -141,7 +150,7 @@ public final class ExpressionOperator extends ExpressionItem {
      */
     public static class Builder
             extends ExpressionItem.Builder<ExpressionOperator, Builder> {
-        private final Op op;
+        private Op op;
 
         private List<ExpressionItem> children = new ArrayList<>();
 
@@ -165,8 +174,8 @@ public final class ExpressionOperator extends ExpressionItem {
         /**
          * @param value Set the logical operator to apply to all the children items
          */
-        public Builder(final Op value) {
-            this.op = value;
+        public Builder(final Op op) {
+            op(op);
         }
 
         /**
@@ -177,7 +186,19 @@ public final class ExpressionOperator extends ExpressionItem {
          */
         public Builder(final Boolean enabled, final Op op) {
             super(enabled);
+            op(op);
+        }
+
+        /**
+         * Changes the operator of this builder
+         *
+         * @param op The operator to set for this builder
+         * @return The {@link Builder}, enabling method chaining
+         */
+        public Builder op(final Op op) {
+            Objects.requireNonNull(op, "Attempt to set null operation");
             this.op = op;
+            return this;
         }
 
         /**
@@ -246,7 +267,7 @@ public final class ExpressionOperator extends ExpressionItem {
          * A convenience function for adding terms in one go, the parameters should read fairly clearly
          *
          * @param field     The field name
-         * @param condition The condition to apply to the valud
+         * @param condition The condition to apply to the value
          * @param value     The value
          * @return this builder, with the completed term added.
          */
@@ -265,7 +286,7 @@ public final class ExpressionOperator extends ExpressionItem {
          * A convenience function for adding dictionary terms in one go, the parameters should read fairly clearly
          *
          * @param field      The field name
-         * @param condition  The condition to apply to the valud
+         * @param condition  The condition to apply to the value
          * @param dictionary The dictionary
          * @return this builder, with the completed term added.
          */